+++ src/tracer/src/cli/handlers/test/mod.rs
@@ -2,5 +2,6 @@ pub mod arguments;
 pub mod git;
 mod handler;
 pub mod pipeline;
+pub mod pixi;
 
 pub use handler::test;
